pyproject.toml
[project]
name = "aegis"
version = "0.1.0"
description = "A sovereign, local-first, hardware-aware autonomous agent framework."
authors = [{ name = "Henry Williams", email = "henry.williams@example.com" }]
requires-python = ">=3.11.9"

dependencies = [
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "py-cpuinfo>=9.0.0",
    "psutil>=5.9.0",
    "nvidia-ml-py>=12.0.0",
    "openvino-dev>=2023.3.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0"
]
